🐛 Fix erronous cors handling in proxy
Turns out, the cors() middleware rejects all requests that doesn't validate with a 400 response. Therefore there is no need to validate the origin in the proxy.

If there is no origin (meaning machine to machine), it doesn't matter what the cors headers are.
Else, we clear what ever comes from the backend and replace it with the request's origin.