✅ Add support for multiple receivers of token cookies
This need arose due to having a SSR setup. On the initial request in an SSR context, the SSR server will require the tokens so that it can proxy them on to the Gatekeeper. Solution is to allow custom cookie domains. Limitation is that the Gatekeeper and the SSR server must be on the same top level domain